GOOD-29_Sprint2 started testing

+++ app/src/main/java/com/example/habittracker/MainScreen.kt
@@ -135,7 +135,7 @@ fun EditWindow(onEvent: (HabitEvent) -> Unit, state: HabitState){
         CustomTextField(
             value = state.editFreq.toString(),
             label = "Frequency:",
-            onchange = { onEvent(HabitEvent.UpDateEditFreq(it.toInt())) },
+            onchange = { onEvent(HabitEvent.UpDateEditFreq(it)) },
             manager = focusManager)
